move prisma to server/utils
