+++ server/api/experiments/attributes/[id].get.ts
@@ -1,4 +1,4 @@
-import prisma from "~~/lib/prisma"
+import prisma from "~~/server/utils/prisma"
 
 export default defineEventHandler(async (event) => {
   const id = getRouterParam(event, "id")
