+++ server/api/legal/[slug].get.ts
@@ -1,5 +1,5 @@
 import { validate as uuidValidate } from "uuid"
-import prisma from "~~/lib/prisma"
+import prisma from "~~/server/utils/prisma"
 
 export default defineEventHandler(async (event) => {
   const slugOrId = getRouterParam(event, "slug")
