+++ server/api/legal/index.get.ts
@@ -1,4 +1,4 @@
-import prisma from "~~/lib/prisma"
+import prisma from "~~/server/utils/prisma"
 
 export default defineEventHandler(async () => {
   const documents = await prisma.legalDocument.findMany()
